+++ frontend/src/app/services/class/classService.tsx
@@ -2,15 +2,18 @@ export const callGetAllClasses = async (
   bucketName: string,
   databaseName: string
 ) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const response = await fetch(
-    "http://localhost:8000/classes/get-all-classes/" +
-      bucketName +
-      "/" +
-      databaseName,
+    `http://localhost:8000/classes/get-all-classes/${bucketName}/${databaseName}`,
     {
       method: "GET",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     }
   );
@@ -22,22 +25,64 @@ export const callGetAllClasses = async (
   return await response.json();
 };
 
+export const callDownloadClass = async (
+  bucketName: string,
+  databaseName: string,
+  className: string
+) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
+  const response = await fetch(
+    `http://localhost:8000/classes/download-class/${bucketName}/${databaseName}/${className}`,
+    {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    }
+  );
+
+  if (!response.ok) {
+    throw new Error("Failed to download class");
+  }
+
+  const blob = await response.blob();
+
+  const downloadUrl = window.URL.createObjectURL(blob);
+
+  // Temporary download link
+  const a = document.createElement("a");
+  a.href = downloadUrl;
+  a.download = `${className}.zip`;
+  document.body.appendChild(a);
+
+  a.click();
+
+  document.body.removeChild(a);
+
+  window.URL.revokeObjectURL(downloadUrl);
+};
+
 export const callRemoveClass = async (
   bucketName: string,
   databaseName: string,
   className: string
 ) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const response = await fetch(
-    "http://localhost:8000/classes/delete-class/" +
-      bucketName +
-      "/" +
-      databaseName +
-      "/" +
-      className,
+    `http://localhost:8000/classes/delete-class/${bucketName}/${databaseName}/${className}`,
     {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     }
   );
